+++ conans/conan_server.py
@@ -1,8 +1,15 @@
-from conans.server.server_launcher import main
+import argparse
+
+from conans.server.launcher import ServerLauncher
 
 
 def run():
-    main()
+    parser = argparse.ArgumentParser(description='Launch the server')
+    parser.add_argument('--migrate', default=False, action='store_true',
+                        help='Run the pending migrations')
+    args = parser.parse_args()
+    launcher = ServerLauncher(force_migration=args.migrate)
+    launcher.launch()
 
 
 if __name__ == '__main__':
